Break out useExternalAuth hook
We will use this on the tasks page.
